# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ShopLab API
  version: 0.1.0
  summary: The JSON API behind the ShopLab demo store.
  description: |
    ShopLab is a demo online store built as a test target. Everything the web UI does goes through this API.

    **Demo site. No real payments or personal data.** Data can be wiped at any time (`POST /api/test/reset`).

    ## Conventions
    - Errors always use one shape: `{"error": {"code", "message", "fieldErrors"?}}` (schema `Error`). `fieldErrors` maps a field name to a message.
    - Lists use `{"data": [], "page", "pageSize", "total"}`. A page past the end is `200` with an empty `data` and the real `total`.
    - Money is always an integer number of cents (`priceCents`, `totalCents`, ...). Only the product list price filters (`minPrice`, `maxPrice`) are in dollars.
    - Timestamps are ISO-8601 UTC strings. Dates are `YYYY-MM-DD`.
    - Status codes: 400 validation, 401 not logged in, 402 payment declined, 403 forbidden, 404 not found, 409 conflict (such as insufficient stock), 413 upload too large, 423 locked account, 429 rate limited.

    ## Authentication
    `POST /api/auth/login` sets an httpOnly cookie (`shoplab_session`) and also returns the same value as `token`. Send either the cookie or `Authorization: Bearer <token>`; if both are sent the bearer token is used.
    Endpoints under `/api/admin/*` need an admin account: `401` when not logged in, `403` for a customer.
...

...

...
  license:
    name: MIT
servers:
  - url: /
...
security: []
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Create an account
      description: Creates a customer account, logs it in (sets the session cookie) and returns the token.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Account created and logged in.
          headers:
            Set-Cookie: { $ref: '#/components/headers/SetSessionCookie' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSession' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Log in
      description: |
        Sets an httpOnly cookie and returns the token. A wrong password or unknown email is `401 INVALID_CREDENTIALS`.
        The locked account answers `423 ACCOUNT_LOCKED` only when the password is right. After 5 failed attempts within
        60 seconds for the same client and email the next attempt is `429` with a `Retry-After` header.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Logged in.
          headers:
            Set-Cookie: { $ref: '#/components/headers/SetSessionCookie' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSession' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/InvalidCredentials' }
        '423':
          description: The account is locked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Too many failed attempts.
          headers:
            Retry-After:
              description: Seconds to wait before trying again.
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: End the session
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '204': { description: Logged out. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /api/auth/me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: The current user
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: The logged-in user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    patch:
      tags: [Auth]
      operationId: updateMe
      summary: Change the display name
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 2, maxLength: 60 }
      responses:
        '200':
          description: The updated user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /api/cart:
    get:
      tags: [Cart]
      operationId: getCart
      summary: The current cart with totals
      description: Totals use Standard shipping; the method is chosen at checkout.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: The cart.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    delete:
      tags: [Cart]
      operationId: clearCart
      summary: Clear the cart and its coupon
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: The (now empty) cart.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /api/cart/items:
    post:
      tags: [Cart]
      operationId: addCartItem
      summary: Add a product to the cart
      description: |
        `variantId` is required for products that have variants and not allowed for products that do not. Quantity is 1 to 10 and a line
        can never hold more than the stock or 10 (409 with the reason). An unknown or inactive product is a 404. 201 for a new line,
        200 when it merged into an existing line.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, quantity]
              properties:
                productId: { type: integer, minimum: 1 }
                variantId: { type: [integer, 'null'], minimum: 1 }
                quantity: { type: integer, minimum: 1, maximum: 10 }
      responses:
        '200':
          description: Merged into an existing line. Returns the whole cart.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
        '201':
          description: A new line was added. Returns the whole cart.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/cart/items/{itemId}:
    patch:
      tags: [Cart]
      operationId: updateCartItem
      summary: Change a line's quantity
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ItemId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: { type: integer, minimum: 1, maximum: 10 }
      responses:
        '200':
          description: The cart.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Cart]
      operationId: removeCartItem
      summary: Remove a line
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ItemId' }
      responses:
        '200':
          description: The cart.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/cart/coupon:
    post:
      tags: [Cart]
      operationId: applyCoupon
      summary: Apply a coupon
      description: |
        One coupon at a time (a second, or the same again, is a 409). Codes are matched ignoring case. 400 when the code is unknown,
        the cart is empty, or the coupon is expired, already used by this account, or below its minimum subtotal.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, examples: [SAVE10] }
      responses:
        '200':
          description: The cart with the coupon.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Cart]
      operationId: removeCoupon
      summary: Remove the coupon
      description: 200 even when there was none.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: The cart.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /api/cart/merge:
    post:
      tags: [Cart]
      operationId: mergeGuestCart
      summary: Merge a guest cart into the server cart
      description: Called by the UI right after login. Quantities add up and are capped at min(10, stock); unavailable products are skipped.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  maxItems: 100
                  items:
                    type: object
                    required: [productId, quantity]
                    properties:
                      productId: { type: integer, minimum: 1 }
                      variantId: { type: [integer, 'null'], minimum: 1 }
                      quantity: { type: integer, minimum: 1, maximum: 10 }
      responses:
        '200':
          description: The merged cart and what was changed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MergeResult' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: shoplab_session
      description: The httpOnly session cookie set by login or register.
    bearerAuth:
      type: http
      scheme: bearer
      description: The `token` returned by login or register.
    testKey:
      type: apiKey
      in: header
      name: X-Test-Key
      description: Required for /api/test/* when the server has TEST_API_KEY set.
  headers:
    SetSessionCookie:
      description: The httpOnly, SameSite=Lax `shoplab_session` cookie (Secure over HTTPS).
      schema: { type: string }
  parameters:
...
    ItemId:
      name: itemId
      in: path
      required: true
      description: A cart line ID.
      schema: { type: integer, minimum: 1 }
...
  responses:
    ValidationError:
      description: Validation failed. `fieldErrors` lists every problem.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthenticated:
      description: Not logged in (or, for /api/test/*, a missing or wrong X-Test-Key).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InvalidCredentials:
      description: Wrong email or password (the same message for both).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: The request conflicts with the current state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHENTICATED
        - INVALID_CREDENTIALS
        - PAYMENT_DECLINED
        - FORBIDDEN
        - PAYLOAD_TOO_LARGE
        - NOT_FOUND
        - CONFLICT
        - ACCOUNT_LOCKED
        - RATE_LIMITED
        - SERVICE_UNAVAILABLE
        - INTERNAL_ERROR
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message: { type: string }
            fieldErrors:
              type: object
              description: Field name to message. Keys can be nested names such as `address.postalCode` or `items.<cartItemId>`.
              additionalProperties: { type: string }
...
    User:
      type: object
      required: [id, name, email, role]
      additionalProperties: false
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [customer, admin] }
    UserResponse:
      type: object
      required: [user]
      additionalProperties: false
      properties:
        user: { $ref: '#/components/schemas/User' }
    AuthSession:
      type: object
      required: [token, user, expiresAt]
      additionalProperties: false
      properties:
        token: { type: string, description: The same value as the session cookie. }
        user: { $ref: '#/components/schemas/User' }
        expiresAt: { type: string, format: date-time }
    RegisterRequest:
      type: object
      required: [name, email, password, confirmPassword, acceptTerms]
      properties:
        name: { type: string, minLength: 2, maxLength: 60 }
        email: { type: string, maxLength: 254 }
        password: { type: string, description: At least 8 characters with an uppercase letter, a digit and a symbol. }
        confirmPassword: { type: string }
        acceptTerms: { type: boolean, const: true }
        rememberMe: { type: boolean, description: Session lasts 7 days instead of 1 hour. }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
        rememberMe: { type: boolean }
...
    ShippingMethod:
      type: string
      enum: [standard, express]
    CartItem:
      type: object
      required: [id, productId, variantId, name, category, imageCount, imagePath, variantLabel, unitPriceCents, regularPriceCents, onSale, quantity, lineTotalCents, stock, maxQuantity, inStock]
      additionalProperties: false
      properties:
        id: { type: integer }
        productId: { type: integer }
        variantId: { type: [integer, 'null'] }
        name: { type: string }
        category: { type: string }
        imageCount: { type: integer }
        imagePath: { type: [string, 'null'] }
        variantLabel: { type: [string, 'null'] }
        unitPriceCents: { type: integer }
        regularPriceCents: { type: integer }
        onSale: { type: boolean }
        quantity: { type: integer, minimum: 1, maximum: 10 }
        lineTotalCents: { type: integer }
        stock: { type: integer, minimum: 0, description: Units available now (0 for an inactive product). }
        maxQuantity: { type: integer, minimum: 0 }
        inStock: { type: boolean }
    CartCoupon:
      type: object
      required: [code, description, applied, reason, message]
      additionalProperties: false
      properties:
        code: { type: string }
        description: { type: string }
        applied: { type: boolean, description: False when the coupon is attached but does not qualify right now. }
        reason:
          type: [string, 'null']
          enum: [expired, below_minimum, already_used, null]
        message: { type: [string, 'null'] }
    CartTotals:
      type: object
      required: [subtotalCents, discountCents, shippingMethod, shippingCents, taxCents, totalCents]
      additionalProperties: false
      properties:
        subtotalCents: { type: integer }
        discountCents: { type: integer }
        shippingMethod: { $ref: '#/components/schemas/ShippingMethod' }
        shippingCents: { type: integer }
        taxCents: { type: integer }
        totalCents: { type: integer }
    Cart:
      allOf:
        - $ref: '#/components/schemas/CartFields'
      unevaluatedProperties: false
    CartFields:
      type: object
      description: The cart. Used on its own as Cart and extended by Quote.
      required: [items, itemCount, coupon, totals]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CartItem' }
        itemCount: { type: integer, minimum: 0, description: Sum of the quantities (the header badge). }
        coupon:
          oneOf:
            - $ref: '#/components/schemas/CartCoupon'
            - type: 'null'
        totals: { $ref: '#/components/schemas/CartTotals' }
    MergeAdjustment:
      type: object
      required: [productId, variantId, name, requested, resulting, reason]
      additionalProperties: false
      properties:
        productId: { type: integer }
        variantId: { type: [integer, 'null'] }
        name: { type: string }
        requested: { type: integer }
        resulting: { type: integer }
        reason: { type: string, enum: [unavailable, out_of_stock, capped] }
    MergeResult:
      type: object
      required: [cart, adjustments]
      additionalProperties: false
      properties:
        cart: { $ref: '#/components/schemas/Cart' }
        adjustments:
          type: array
          items: { $ref: '#/components/schemas/MergeAdjustment' }
...